+++ CHANGES.md
@@ -4,6 +4,7 @@
 * Used the line number colour for the print margin.
 * Made PHP variables readable in themes that do not colour local
   variables.
+* Coloured the debug instruction pointer.
 
 0.7.0 (2011-01-29)
 ==================